pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sqlean-stubs"
version = "3.50.4"
description = "Type hints for sqlean.py (sqlite3 + extensions)"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Zlib"}
authors = [
    {name = "Kracekumar", email = "me@kracekumar.com"},
]
keywords = ["sqlite", "sqlean", "type-hints", "stubs", "type-checking"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Zlib License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Typing :: Typed",
]

[project.urls]
Repository = "https://github.com/nalgeon/sqlean.py"
Documentation = "https://github.com/nalgeon/sqlean"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "mypy>=1.0",
    "ty>=0.0.7",
    "tox>=4.0",
    "sqlean.py>=3.50",
]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "mypy>=1.0",
    "ty>=0.0.7",
    "tox>=4.0",
    "sqlean.py>=3.50",
]

[tool.setuptools]
packages = ["sqlean"]

[tool.setuptools.package-data]
sqlean = ["py.typed"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.ty.environment]
python-version = "3.9"

[tool.ty.src]
exclude = ["setup.py"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v"
